Change polling station number from i64 to u32

+++ backend/src/audit_log/audit_event.rs
@@ -74,7 +74,7 @@ pub struct PollingStationDetails {
     pub polling_station_id: u32,
     pub polling_station_election_id: u32,
     pub polling_station_name: String,
-    pub polling_station_number: i64,
+    pub polling_station_number: u32,
     pub polling_station_number_of_voters: Option<i64>,
     #[serde(skip_serializing_if = "Option::is_none")]
     #[schema(nullable = false)]
